Working on authentication
App receives:
POST requests to /auth/signup and /auth/signin with an email (sanitized) and password as form elements, replying with a JWT that is set to expire after 15 mins. This will hold the user session só there's no constant login required.
A GET request to /users/me with the JWT received from one of those POST requests as a request header "Authorization": "Bearer <JWT>" to retrieve user info